requirements.txt
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
matplotlib==3.9.0
numpy==1.26.4
reportlab==4.1.0
